refactor: reorganize projects page code structure
- Extract components into separate files for better maintainability
- Create custom hooks for filter logic (useProjectFilters)
- Separate data layer (projects.ts) from presentation
- Add utility functions for skill icons and helpers
- Implement barrel exports for cleaner imports
- Add comprehensive TypeScript interfaces
- Create detailed refactoring documentation

Components created:
- ProjectFilters: filter controls component
- ProjectGrid: project grid with empty state
- ProjectCard: individual project card
- ProjectSkillsSidebar: left sidebar with skills display
- ProjectDetailsSidebar: right sidebar with project details

Benefits:
- Improved code organization and readability
- Better separation of concerns
- Enhanced reusability and testability
- Easier maintenance and debugging
- Consistent component architecture
