+++ src/components/index.ts
@@ -1,0 +1,11 @@
+// Componentes principais
+export { default as Navigation } from './Navigation';
+export { default as Footer } from './Footer';
+export { AnimatedElement } from './AnimatedElement';
+
+// Componentes específicos de projetos
+export { ProjectFilters } from './ProjectFilters';
+export { ProjectGrid } from './ProjectGrid';
+export { ProjectCard } from './ProjectCard';
+export { ProjectSkillsSidebar } from './ProjectSkillsSidebar';
+export { ProjectDetailsSidebar } from './ProjectDetailsSidebar';
