+++ src/data/index.ts
@@ -1,0 +1,3 @@
+// Data exports
+export { projects } from './projects';
+export type { Project } from './projects';
