+++ src/hooks/index.ts
@@ -1,0 +1,2 @@
+// Custom hooks
+export { useProjectFilters } from './useProjectFilters';
