+++ src/utils/index.ts
@@ -1,0 +1,2 @@
+// Utility functions
+export { getSkillIcon } from './skillIcons';
